Chore : 관습적인 서비스 인터페이스 삭제
